parser update

+++ include/server.h
@@ -2,7 +2,7 @@
 #define SERVER_H
 
 
-int check_port(int port);
+int check_port(unsigned short port);
